requirements.txt
# Python dependencies
# Any changes in these package requirements
# should be reflected in setup.py as well.
mysqlclient
evelink
dnspython
passlib
requests>=2.9.1
bcrypt
python-slugify>=1.2
requests-oauthlib

redis
celery>=4.0.2

# Django Stuff #
django>=1.10,<2.0
django-bootstrap-form
django-bootstrap-pagination
django-registration
django-sortedm2m
django-redis-cache>=1.7.1
django-recaptcha
django-celery-beat

# Openfire
sleekxmpp
openfire-restapi

adarnauth-esi<2.0
